test(database): MySQL 統合テストに skip ガードを追加 (#1405)
DatabaseMySQL スイート（opt-in 統合テスト）は MySQL 未起動時に接続を 20 回
リトライ後 self::fail() しており、setUp が毎メソッド走るため計 13 回 × 5 秒
ハングしてから失敗していた。CLAUDE.md の FT ルール「MySQL 未設定なら自動 skip」
と乖離していた。

- 共有 trait Nene2\Tests\Database\MySql\RequiresMySql を新設し、3 ファイルに
  完全重複していた connection()/config()/env()（各 ~36 行）を集約。
- 到達不能時は self::fail() ではなく self::markTestSkipped()。MySQL 起動時は
  従来どおり実行（20 回リトライで起動レースも吸収）。
- 到達不能判定を private static にキャッシュし、初回失敗後の setUp は即スキップ。
- compose.yaml が DB_HOST を常設するため env 有無では判定できず、接続試行→
  到達不能なら skip という設計にした。

検証: composer cs / analyse(PHPStan L8) 通過、DatabaseMySQL を MySQL 無しで
実行 → 13 件全 skip・exit 0、デフォルト test(NENE2,Database) 456 件全通過。
